rename folder `api/cpu` in `api/host`
