+++ include/alpaka/api/cpu.hpp
@@ -4,8 +4,8 @@
 
 #pragma once
 
-#include "alpaka/api/cpu/Api.hpp"
-#include "alpaka/api/cpu/Device.hpp"
-#include "alpaka/api/cpu/Platform.hpp"
-#include "alpaka/api/cpu/Queue.hpp"
-#include "alpaka/api/cpu/atomic.hpp"
+#include "alpaka/api/host/Api.hpp"
+#include "alpaka/api/host/Device.hpp"
+#include "alpaka/api/host/Platform.hpp"
+#include "alpaka/api/host/Queue.hpp"
+#include "alpaka/api/host/atomic.hpp"
