+++ include/alpaka/api/executor.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/api/cpu/executor.hpp"
+#include "alpaka/api/host/executor.hpp"
 #include "alpaka/api/cuda/executor.hpp"
 #include "alpaka/api/hip/executor.hpp"
 #include "alpaka/api/oneApi/executor.hpp"
