+++ include/alpaka/api/host/IdxLayer.hpp
@@ -61,5 +61,5 @@ namespace alpaka::onAcc
             T_Idx m_idx;
             T_Count m_count;
         };
-    } // namespace cpu
+    } // namespace host
 } // namespace alpaka::onAcc
