+++ include/alpaka/api/host/block/mem/SingleThreadStaticShared.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/api/cpu/block/mem/SharedStorage.hpp"
+#include "alpaka/api/host/block/mem/SharedStorage.hpp"
 #include "alpaka/core/Vectorize.hpp"
 #include "alpaka/core/common.hpp"
 
@@ -51,5 +51,5 @@ namespace alpaka::onAcc
             {
             }
         };
-    } // namespace cpu
+    } // namespace host
 } // namespace alpaka::onAcc
