+++ include/alpaka/api/host/cpuArchSize.hpp
