+++ include/alpaka/api/host/exec/Serial.hpp
@@ -5,9 +5,9 @@
 #pragma once
 
 #include "alpaka/Vec.hpp"
-#include "alpaka/api/cpu/IdxLayer.hpp"
-#include "alpaka/api/cpu/block/mem/SingleThreadStaticShared.hpp"
-#include "alpaka/api/cpu/block/sync/NoOp.hpp"
+#include "alpaka/api/host/IdxLayer.hpp"
+#include "alpaka/api/host/block/mem/SingleThreadStaticShared.hpp"
+#include "alpaka/api/host/block/sync/NoOp.hpp"
 #include "alpaka/core/Dict.hpp"
 #include "alpaka/meta/NdLoop.hpp"
 #include "alpaka/onAcc/Acc.hpp"
@@ -80,7 +80,7 @@ namespace alpaka::onHost
 
             ThreadSpecType m_threadBlocking;
         };
-    } // namespace cpu
+    } // namespace host
 
     inline auto makeAcc(exec::CpuSerial, auto const& threadBlocking)
     {
