+++ include/alpaka/api/oneApi/Api.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/api/cpu/cpuArchSize.hpp"
+#include "alpaka/api/host/cpuArchSize.hpp"
 #include "alpaka/api/syclGeneric/Api.hpp"
 #include "alpaka/api/trait.hpp"
 #include "alpaka/concepts.hpp"
